backend/requirements.txt
Django==5.0
djangorestframework>=3.15
pymysql==1.1.0
python-dotenv==1.0.0
django-cors-headers==4.3.0
pytest==7.4.0
cryptography==41.0.7
Pillow==10.1.0